pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mempalace-mcp-server"
version = "0.1.0"
description = "Streamable-HTTP MCP server for MemPalace palaces — single-writer, multi-device, per-palace"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "mcp==1.27.0",
    "chromadb==1.5.8",
    "uvicorn==0.45.0",
    "starlette>=0.47",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "typer>=0.12",
]

[project.scripts]
mempalace-server = "cli.__main__:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "cli*"]
